Update action path to work on windows
Previously the path to the action script was `${{ github.action_path }}`
which uses the native path format. On Windows machines this ends up
passing paths with `\` to bash, which bash tries to interpret as escape
characters. "c:\a\b" would get be interpreted as "c:ab".

Now the path replaces all instances of `\` with `/` to work with bash on
Windows. This assumes that `\` would never be used on other github
runners.

Solution was copied from https://github.com/actions/runner/issues/1066